+++ src/num_traits/mod.rs
@@ -1,12 +1,15 @@
 mod cast;
 mod checked;
 mod containing_float;
+mod convert;
 mod float;
 mod goniometric;
 mod infinity;
 mod into_float;
 mod isqrt;
 mod large_type;
+mod limits;
+mod neg_infinity;
 mod normal_limits;
 mod one;
 mod saturating;
@@ -14,13 +17,10 @@ mod scale;
 mod sqrt;
 mod two;
 mod zero;
-mod neg_infinity;
-mod limits;
-mod convert;
 
 pub use self::{
-    cast::*, checked::*, containing_float::*, float::*, goniometric::*,
-    infinity::*, into_float::*, isqrt::*, large_type::*, normal_limits::*,
-    one::*, saturating::*, scale::*, sqrt::*, two::*, zero::*, neg_infinity::*,
-    limits::*, convert::*,
+    cast::*, checked::*, containing_float::*, convert::*, float::*,
+    goniometric::*, infinity::*, into_float::*, isqrt::*, large_type::*,
+    limits::*, neg_infinity::*, normal_limits::*, one::*, saturating::*,
+    scale::*, sqrt::*, two::*, zero::*,
 };