+++ src/num_traits/two.rs
@@ -2,7 +2,16 @@ pub trait Two {
     const TWO: Self;
 }
 
+#[macro_export]
 macro_rules! impl_two {
+    ($t:ident < $g:ident >) => {
+        impl<$g: $crate::Two + Copy> $crate::Two for $t<$g> {
+            const TWO: Self = Self::same_components($g::TWO);
+        }
+    };
+}
+
+macro_rules! impl_base_two {
     ($($t:ident),* -> $v:expr) => {
         $(impl Two for $t {
             const TWO: Self = $v;
@@ -10,5 +19,5 @@ macro_rules! impl_two {
     };
 }
 
-impl_two!(u8, i8, u16, i16, u32, i32, u64, i64, usize, isize, u128, i128 -> 2);
-impl_two!(f32, f64 -> 2.);
+impl_base_two!(u8, i8, u16, i16, u32, i32, u64, i64, usize, isize, u128, i128 -> 2);
+impl_base_two!(f32, f64 -> 2.);