+++ src/num_traits/zero.rs
@@ -2,7 +2,16 @@ pub trait Zero {
     const ZERO: Self;
 }
 
-macro_rules! impl_one {
+#[macro_export]
+macro_rules! impl_zero {
+    ($t:ident < $g:ident >) => {
+        impl<$g: $crate::Zero + Copy> $crate::Zero for $t<$g> {
+            const ZERO: Self = Self::same_components($g::ZERO);
+        }
+    };
+}
+
+macro_rules! impl_base_zero {
     ($($t:ident),* -> $v:expr) => {
         $(impl Zero for $t {
             const ZERO: Self = $v;
@@ -10,5 +19,5 @@ macro_rules! impl_one {
     };
 }
 
-impl_one!(u8, i8, u16, i16, u32, i32, u64, i64, usize, isize, u128, i128 -> 0);
-impl_one!(f32, f64 -> 0.);
+impl_base_zero!(u8, i8, u16, i16, u32, i32, u64, i64, usize, isize, u128, i128 -> 0);
+impl_base_zero!(f32, f64 -> 0.);