+++ src/traits/mod.rs
@@ -1,9 +1,9 @@
 mod comp_arithm;
 mod map_ext;
+mod map_traits;
 mod range_ext;
 mod rect_ext;
 mod two_component;
-mod map_traits;
 
 pub use self::{
     comp_arithm::*, map_ext::*, range_ext::*, rect_ext::*, two_component::*,