+++ src/vec2.rs
@@ -9,7 +9,8 @@ use std::{
 };
 
 use crate::{
-    Cast, Checked, Convert, Float, Goniometric, Infinity, IntoFloat, Isqrt, LargeType, MapExt, NegInfinity, NormalLimits, One, Scale, Sqrt, Two, Vec2RangeIter, Zero
+    Checked, Float, Goniometric, IntoFloat, Isqrt, LargeType, MapExt,
+    NormalLimits, Scale, Sqrt, Vec2RangeIter, Zero, impl_traits,
 };
 
 /// Represents two dimensional vector. Can be used as vector, point, size or
@@ -28,6 +29,14 @@ impl<T> Vec2<T> {
     /// Creates new two dimensional vector from its components.
     pub const fn new(x: T, y: T) -> Self {
         Self { x, y }
+    }
+
+    /// Set all components to the same value.
+    pub const fn same_components(v: T) -> Self
+    where
+        T: Copy,
+    {
+        Self::new(v, v)
     }
 
     /// Get the width. Alias to the first coordinate (x, [0])
@@ -620,47 +629,7 @@ impl<T> Vec2<&T> {
     }
 }
 
-impl<T: Zero> Zero for Vec2<T> {
-    /// 2D vectors with all components set to zero.
-    const ZERO: Vec2<T> = Vec2 {
-        x: T::ZERO,
-        y: T::ZERO,
-    };
-}
-
-impl<T: One> One for Vec2<T> {
-    const ONE: Self = Vec2::new(T::ONE, T::ONE);
-}
-
-impl<T: Two> Two for Vec2<T> {
-    const TWO: Self = Vec2::new(T::TWO, T::TWO);
-}
-
-impl<T: Infinity> Infinity for Vec2<T> {
-    const INFINITY: Self = Vec2::new(T::INFINITY, T::INFINITY);
-}
-
-impl<T: NegInfinity> NegInfinity for Vec2<T> {
-    const NEG_INFINITY: Self = Vec2::new(T::NEG_INFINITY, T::NEG_INFINITY);
-}
-
-impl<R, T: Cast<R>> Cast<Vec2<R>> for Vec2<T> {
-    fn cast(self) -> Vec2<R> {
-        self.map(T::cast)
-    }
-}
-
-impl<R, T: Scale<R>> Scale<Vec2<R>> for Vec2<T> {
-    fn scale(self) -> Vec2<R> {
-        self.map(T::scale)
-    }
-}
-
-impl<R, T: Convert<R>> Convert<Vec2<R>> for Vec2<T> {
-    fn convert(self) -> Vec2<R> {
-        self.map(T::convert)
-    }
-}
+impl_traits!(Vec2<T> => VecTraits);
 
 impl<T> From<(T, T)> for Vec2<T> {
     fn from((x, y): (T, T)) -> Self {